pkg/op-app/pyproject.toml
[tool.poetry]
name = "op_app"
version = "0.1.0"
description = ""
authors = ["José Ribeiro <me@joseribeiro.dev>"]

[tool.poetry.dependencies]
python = "^3.12"
op-core = { path = "../op-core", develop = true }
op-brains = { path = "../op-brains", develop = true }
cow-brains = { path = "../cow-brains", develop = true }
op-data = { path = "../op-data", develop = true }
op-artifacts = { path = "../op-artifacts", develop = true }
python-dotenv = "^1.0.1"
quart = "^0.19.6"
quart-cors = "^0.7.0"
uvicorn = "^0.30.6"
aiofiles = "^24.1.0"
sentence-transformers = "^5.2.3"
groq = "^0.11.0"
boto3 = "^1.35.27"
google-generativeai = "^0.8.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
